requirements.txt
openai
azure-cognitiveservices-speech
python-dotenv
pydub